--- recommender_systems_v2/requirements.txt
@@ -1,10 +1,0 @@
-flask==3.0.0
-elasticsearch==8.17.0
-sentence-transformers==3.3.1
-torch
-numpy
-scikit-learn
-redis==5.2.1
-pandas
-openai
-requests
